Metal mine upgrade (#4)
Upgrading a metal mine leads to an increased metal income. One can start a metal mine upgrade from the newly implemented planet's page, where the conditions are displayed, such as the cost and time to complete the upgrade.

Upgrading requires resources and time. The `MetalMineUpgrade` model supports the calculations of cost and duration for a given level upgrade. Cost and duration increase exponentially level after level.

Only one upgrade can be started at a time.
If an upgrade is rejected (missing resources, another one still in progress), the error is displayed to the user as a flash message.

Once the upgrade is started, it's not possible to start another upgrade. The pending upgrade's remaining time is showed and the cost of the upgrade is deducted from the planet's resources.

A background job is scheduled for the end time, so that the completion logic is immediately executed. Completing an upgrade marks it as finished, update the metal mine's level and update the planet's resources so that the last update already uses the new income.