Metal mine upgrade remaining time as a dynamic countdown
Before the remaining time of a metal mine upgrade was displayed as
human-readable works, but the user had to refresh the page to see the
up-to-date value.

Now, the remaining time is displayed in a more numeric format. Every
second, it decreases by a second to show the actual remaining time,
without having to refresh the page.

Co-authored-by: Jose Blanco <jose.blanco@thoughtbot.com>